refactor(lo,nco): one scalar LO, and a control port that is double everywhere
Two changes in one commit because the first is the second's prerequisite,
not because they are one idea.

DELETE THE AVX-512 KERNELS. lo_core.c carried two, each with a complete
scalar fallback beside it. `-march=x86-64-v2` is the shipped baseline, so
`__AVX512F__` was never defined in any wheel, any CI job, or any local build
that did not opt into DOPPLER_NATIVE -- neither kernel was ever compiled,
let alone tested. And where one did build, it disagreed: `lo_steps_ctrl`'s
vector path folded its control in float32 via _mm512_cvtps_epu32 while the
scalar path folds in double through the shared primitive, so the same object
would have produced different phase increments depending on the CPU. Two
implementations of one function, one unreachable and the two disagreeing, is
strictly worse than one. This retires the corresponding entry from
scripts/.phase-conversion-allow, so the ratchet shrinks 8 -> 7.

THE CONTROL PORT IS NOW `double`. Four signatures narrowed to `const float *`
-- nco_steps_u32_ctrl, _scaled_ctrl, _ovf_ctrl and lo_steps_ctrl -- while
every scalar steer site in the library (nco_step_u32*_ctrl, lo_step_ctrl,
symsync, dll) already used `double`, which is also the width the conversion
itself works in. The result was that one requested 0.1 reached the
accumulator as two different phase words depending on which face it entered
by, 7 apart, with nothing documenting which resolution a loop was steering
at. That is the NCO's F2 and, inherited with the port, the LO's F5.

Measured after: the delta is 0, and the dead zone where a control cancels
phase_inc falls from 22.4 ppb to 0.232 ppb against a phase-word LSB of
0.233 -- one LSB, which is the floor no port precision can beat, so F3/F6
stop being defects and become the quantization they always were.

The deletion has to come first: the AVX kernel's 16-lane float prefix scan
cannot take doubles, and rewriting it as an 8-lane double version would have
been effort spent on code no shipped build compiles.

No production C called the narrowing forms -- only the generated bindings
and the C tests -- and the binding still accepts a float32 array by casting,
so no caller breaks.

The two `_ext` fragments carry one hand-restored two-line block, marked and
filed as just-buildit/just-makeit#920: under `pass_capacity`, 0.55.2 allocates
`max_out()` for the fresh-array path and discards the request, which both
truncates above 65536 and mallocs 256 KB per call below it. doppler's #116
large-n tests go red if that is lost again, so it cannot regress silently.

Also here, because the same audit turned them up: lo_step_ctrl's comment
claimed the shared primitive "rounds, not truncates" -- it truncates, and
test_lo_core.c section 21 now pins that -- and nco_steps_u32_ctrl's
docstring called `out=` "essential for a hot per-epoch tracking loop", which
measures at 0-25% below 8192 samples and nothing above it.

+++ native/src/nco/nco_core.c
@@ -186,14 +186,14 @@ nco_steps_u32_ctrl_max_out (nco_state_t *state)
 }
 
 size_t
-nco_steps_u32_ctrl (nco_state_t *state, const float *ctrl, size_t ctrl_len,
+nco_steps_u32_ctrl (nco_state_t *state, const double *ctrl, size_t ctrl_len,
                     uint32_t *out, size_t max_out)
 {
   /* Emission stops at the caller's capacity (jm gh-138). */
   if (ctrl_len > max_out)
     ctrl_len = max_out;
   for (size_t i = 0; i < ctrl_len; i++)
-    out[i] = nco_step_u32_ctrl (state, (double)ctrl[i]);
+    out[i] = nco_step_u32_ctrl (state, ctrl[i]);
   return ctrl_len;
 }
 
@@ -205,14 +205,14 @@ nco_steps_u32_scaled_ctrl_max_out (nco_state_t *state)
 }
 
 size_t
-nco_steps_u32_scaled_ctrl (nco_state_t *state, const float *ctrl,
+nco_steps_u32_scaled_ctrl (nco_state_t *state, const double *ctrl,
                            size_t ctrl_len, uint32_t *out, size_t max_out)
 {
   /* Emission stops at the caller's capacity (jm gh-138). */
   if (ctrl_len > max_out)
     ctrl_len = max_out;
   for (size_t i = 0; i < ctrl_len; i++)
-    out[i] = nco_step_u32_scaled_ctrl (state, (double)ctrl[i]);
+    out[i] = nco_step_u32_scaled_ctrl (state, ctrl[i]);
   return ctrl_len;
 }
 
@@ -224,13 +224,14 @@ nco_steps_u32_ovf_ctrl_max_out (nco_state_t *state)
 }
 
 size_t
-nco_steps_u32_ovf_ctrl (nco_state_t *state, const float *ctrl, size_t ctrl_len,
-                        uint32_t *out, uint8_t *out1, size_t max_out)
+nco_steps_u32_ovf_ctrl (nco_state_t *state, const double *ctrl,
+                        size_t ctrl_len, uint32_t *out, uint8_t *out1,
+                        size_t max_out)
 {
   /* Emission stops at the caller's capacity (jm gh-138). */
   if (ctrl_len > max_out)
     ctrl_len = max_out;
   for (size_t i = 0; i < ctrl_len; i++)
-    out[i] = nco_step_u32_ovf_ctrl (state, (double)ctrl[i], &out1[i]);
+    out[i] = nco_step_u32_ovf_ctrl (state, ctrl[i], &out1[i]);
   return ctrl_len;
 }
